split terrier parts into separate files
