+++ app/frontend/terrier/dropdowns.ts
@@ -3,7 +3,7 @@ import { untypedKey } from "tuff-core/messages"
 import { unique } from "tuff-core/arrays"
 import {PartTag, StatelessPart} from "tuff-core/parts"
 import Overlays from "./overlays"
-import {TerrierPart} from "./parts"
+import TerrierPart from "./parts/terrier-part"
 import Objects from "tuff-core/objects"
 import Theme, {Action, ThemeType} from "./theme"
 import {TerrierApp} from "./app"
