+++ app/frontend/terrier/parts/not-found-page.ts
@@ -1,0 +1,26 @@
+import {Logger} from "tuff-core/logging"
+import Theme, {ThemeType} from "../theme"
+import {TerrierApp} from "../app"
+import PagePart from "./page-part"
+import {NoState, PartTag} from "tuff-core/parts"
+
+const log = new Logger('NotFoundRoute')
+
+/**
+ * Default page part if the router can't find the path.
+ */
+export default class NotFoundRoute<
+    TT extends ThemeType,
+    TApp extends TerrierApp<TT, TApp, TTheme>,
+    TTheme extends Theme<TT>
+> extends PagePart<NoState, TT, TApp, TTheme> {
+    async init() {
+        this.setTitle("Page Not Found")
+    }
+
+    renderContent(parent: PartTag) {
+        log.warn(`Not found: ${this.context.href}`)
+        parent.h1({text: "Not Found"})
+    }
+
+}
